Fixes the flaky runSingle test.
After starting the application the test waits only until all plugins are started. But the
"ok"-test wants to do something while the plugin is running. This needs some time. If
sent the shutdown signal immediately after the plugin was started, the plugin has not
enough time to do it's work. In this case the test runs into the timeout. Now we give
the plugin some time to do it's work before we send the shutdown signal.